Conflicts:
	WowPacketParser/App.config
	WowPacketParser/Enums/ClientVersionBuild.cs
	WowPacketParser/Enums/Version/Opcodes.cs
	WowPacketParser/Enums/Version/UpdateFields.cs
	WowPacketParser/Misc/ClientVersion.cs
	WowPacketParser/SQL/Builders/UnitMisc.cs
	WowPacketParser/WowPacketParser.csproj
	WowPacketParserModule.V5_4_0_17359/Parsers/NpcHandler.cs
	WowPacketParserModule.V6_0_2_19033/Parsers/BattlegroundHandler.cs
	WowPacketParserModule.V6_0_2_19033/Parsers/NpcHandler.cs
	WowPacketParserModule.V7_0_3_22248/Parsers/QuestHandler.cs

+++ WowPacketParserModule.V7_2_0_23758/Parsers/HotfixHandler.cs
@@ -130,7 +130,7 @@ namespace WowPacketParserModule.V7_2_0_23758.Parsers
 
             HotfixData hotfixData = new HotfixData
             {
-                ID = hotfixId,
+                ID = (uint)hotfixId,
                 TableHash = type,
                 RecordID = entry,
                 Deleted = !allow
